requirements.txt
anyio==4.0.0
certifi==2023.7.22
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
polars==0.19.11
psycopg2-binary==2.9.9
Pygments==2.16.1
python-dotenv==1.0.0
rich==13.6.0
sniffio==1.3.0
